create new package for fsspec writers
