+++ unstructured/ingest/runner/writers/fsspec/s3.py
@@ -6,17 +6,19 @@ from unstructured.ingest.interfaces import BaseDestinationConnector
 from unstructured.ingest.runner.writers.base_writer import Writer
 
 if t.TYPE_CHECKING:
-    from unstructured.ingest.connector.gcs import GcsWriteConfig, SimpleGcsConfig
+    from unstructured.ingest.connector.s3 import S3WriteConfig, SimpleS3Config
 
 
 @dataclass
-class GcsWriter(Writer, EnhancedDataClassJsonMixin):
-    connector_config: "SimpleGcsConfig"
-    write_config: "GcsWriteConfig"
+class S3Writer(Writer, EnhancedDataClassJsonMixin):
+    connector_config: "SimpleS3Config"
+    write_config: "S3WriteConfig"
 
     def get_connector(self, **kwargs) -> BaseDestinationConnector:
-        from unstructured.ingest.connector.gcs import GcsDestinationConnector
+        from unstructured.ingest.connector.fsspec.s3 import (
+            S3DestinationConnector,
+        )
 
-        return GcsDestinationConnector(
+        return S3DestinationConnector(
             write_config=self.write_config, connector_config=self.connector_config
         )
